Fix remaining compiler errors for GCC 13 and ACE 7

+++ src/server/game/Achievements/AchievementMgr.h
@@ -59,7 +59,7 @@ enum AchievementCriteriaDataType
 
 enum AchievementCommonCategories
 {  
-    ACHIEVEMENT_CATEOGRY_GENERAL                       = -1,
+    ACHIEVEMENT_CATEOGRY_GENERAL                       = static_cast<uint32>(-1),
     ACHIEVEMENT_CATEGORY_STATISTICS                    =  1
 };
